Mirror unknown anchor login
This renames the undescriptive and somewhat confusing "loginUnknown" to
"loginUnknownAnchor" and splits "login" into "loginKnownAnchor"
and "index.ts".
